made ctx_t and 'uint32_t' to fix kernel<->user conflicts (-1 in user becomes 0x0000ffff in kernel which would expect 0xffffffff)
git-svn-id: http://svn.linux-vserver.org/svn/util-vserver/branches/SYSCALL_SWITCH@394 94cd875c-1c1d-0410-91d2-eb244daf1a30

+++ util-vserver/compat.h
@@ -39,7 +39,8 @@
 #endif
 
 #ifndef HAVE_CTX_T
-typedef short int		ctx_t;
+#include <stdint.h>
+typedef uint32_t		ctx_t;
 #endif
 
 #endif	//  H_UTIL_VSERVER_COMPAT_H
